* changed the TESSreal type to double which suits the PointLine better. * added sln and vcxproj for vs2017

+++ Include/tesselator.h
@@ -129,7 +129,7 @@ enum TessOption
 	TESS_REVERSE_CONTOURS
 };
 
-typedef float TESSreal;
+typedef double TESSreal;
 typedef int TESSindex;
 typedef struct TESStesselator TESStesselator;
 typedef struct TESSalloc TESSalloc;
